refactor: deconstruct ClassroomsDashboard page into useClassroomsDashboard hook and presenter components
